include sheets/ in package
Fixes #68.

Otherwise the tests will fail if run from the installed library, as
then the filename (given relative to the test file) doesn't point to
the requested sheet.

This patch moves sheets/ to src/cssutils/tests/sheets/ and tells
setuptools to include those files in the installation. We can then
access those files in the tests with pkg_resources.