forward createDocument(Type) to minidom
Fixes #69.

We don't want to break other libraries just because we register
cssutils.DOMImplementationCSS as DOMImplementation, without actually
providing the required methods. Applications and libraries depending on
xml.dom.getDOMImplementation() returning a valid implementation will
break once cssutils is imported.

Now, we forward createDocument and createDocumentType to minidom, which
un-breaks those applications, even though the "fix" is not that nice.