Delay compilation of regexes in profile
Fixes #72.

When cssutils was imported, all regular expressions were precompiled,
regardless of whether they were needed or not. This could delay a
"import cssutils" for some seconds. Profiling has shown that most of this
import time is spent in re.compile.

This change moves the compilation of many regexes to when they are first
needed. This is accomplished by the new util.LazyRegex class, which mimics the
interface of native regex objects, but the compilation is only done when the
.search/.match/... method is first called.

Note that
1.) LazyRegexes are currently only used in one place, as profiling has shown
    this as the bottleneck. Excessive use of LazyRegex might cause an overall
    slowdown (So a simple s/re.compile/LazyRegex/ won't fix everything)
2.) This does not mean that cssutils is faster as a whole. Only importing
    cssutils is faster.